Abstracted login_hanlder and login_service away from rest_handler and added validation against username.
